fix(onlinestream): fixed episode being changed when switching providers
refactor(onlinestream): improve code clarity
