+++ seanime-web/src/app/(main)/mediastream/_components/mediastream-video-addons.tsx
@@ -1,20 +1,51 @@
-import { Menu, useAudioOptions } from "@vidstack/react"
+import { __mediastream_autoNextAtom, __mediastream_autoPlayAtom } from "@/app/(main)/mediastream/_lib/mediastream.atoms"
+import { submenuClass, VdsSubmenuButton } from "@/app/(main)/onlinestream/_components/onlinestream-video-addons"
+import { Switch } from "@/components/ui/switch"
+import { Menu } from "@vidstack/react"
+import { useAtom } from "jotai/react"
+import React from "react"
+import { AiFillPlayCircle } from "react-icons/ai"
+import { MdPlaylistPlay } from "react-icons/md"
 
-export function MediastreamAudioSubmenu() {
-    const options = useAudioOptions(),
-        hint = options.selectedTrack?.label
+export function MediastreamPlaybackSubmenu() {
+
+    const [autoPlay, setAutoPlay] = useAtom(__mediastream_autoPlayAtom)
+    const [autoNext, setAutoNext] = useAtom(__mediastream_autoNextAtom)
+
     return (
-        <Menu.Root>
-            <Menu.Button disabled={options.disabled}>Audio ({hint})</Menu.Button>
-            <Menu.Content>
-                <Menu.RadioGroup value={options.selectedValue}>
-                    {options.map(({ label, value, select }) => (
-                        <Menu.Radio value={value} onSelect={select} key={value}>
-                            {label}
-                        </Menu.Radio>
-                    ))}
-                </Menu.RadioGroup>
-            </Menu.Content>
-        </Menu.Root>
+        <>
+            <Menu.Root>
+                <VdsSubmenuButton
+                    label={`Auto Play`}
+                    hint={autoPlay ? "On" : "Off"}
+                    disabled={false}
+                    icon={AiFillPlayCircle}
+                />
+                <Menu.Content className={submenuClass}>
+                    <Switch
+                        label="Auto play"
+                        fieldClass="py-2 px-2"
+                        value={autoPlay}
+                        onValueChange={setAutoPlay}
+                    />
+                </Menu.Content>
+            </Menu.Root>
+            <Menu.Root>
+                <VdsSubmenuButton
+                    label={`Play Next`}
+                    hint={autoNext ? "On" : "Off"}
+                    disabled={false}
+                    icon={MdPlaylistPlay}
+                />
+                <Menu.Content className={submenuClass}>
+                    <Switch
+                        label="Auto play next"
+                        fieldClass="py-2 px-2"
+                        value={autoNext}
+                        onValueChange={setAutoNext}
+                    />
+                </Menu.Content>
+            </Menu.Root>
+        </>
     )
 }
